CertificateVerification/requirement.txt
Flask
web3
PyQRCode
pypng
boto3
requests
python-dotenv
tensorflow
numpy
Pillow
scikit-learn